+++ src/Analyzers/Core/CodeFixes/CodeFixes.projitems
@@ -37,6 +37,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)FileHeaders\AbstractFileHeaderCodeFixProvider.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ForEachCast\AbstractForEachCastCodeFixProvider.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Formatting\FormattingCodeFixProvider.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImplementAbstractClass\AbstractImplementAbstractClassCodeFixProvider.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImplementAbstractClass\ImplementAbstractClassData.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImplementType\ImplementTypeInsertionBehavior.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImplementType\ImplementTypeOptions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImplementType\ImplementTypePropertyGenerationBehavior.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Iterator\AbstractIteratorCodeFixProvider.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MakeFieldReadonly\AbstractMakeFieldReadonlyCodeFixProvider.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MakeMemberStatic\AbstractMakeMemberStaticCodeFixProvider.cs" />
