+++ src/Analyzers/Core/CodeFixes/ImplementAbstractClass/ImplementAbstractClassData.cs
@@ -12,7 +12,7 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeGeneration;
-using Microsoft.CodeAnalysis.Editing;
+
 using Microsoft.CodeAnalysis.ImplementInterface;
 using Microsoft.CodeAnalysis.ImplementType;
 using Microsoft.CodeAnalysis.LanguageService;
@@ -20,6 +20,12 @@ using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Shared.Utilities;
 using Roslyn.Utilities;
+
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Editing;
+#else
+namespace Microsoft.CodeAnalysis.Editing;
+#endif
 
 namespace Microsoft.CodeAnalysis.ImplementAbstractClass;
