+++ src/Analyzers/Core/CodeFixes/ImplementType/ImplementTypePropertyGenerationBehavior.cs
