+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/CodeGeneration/CodeGenerationOptions.cs
@@ -63,40 +63,3 @@ internal readonly record struct CodeAndImportGenerationOptions
         };
 #endif
 }
-
-internal static class CodeGenerationOptionsProviders
-{
-#if !CODE_STYLE
-    public static CodeGenerationOptions GetCodeGenerationOptions(this IOptionsReader options, LanguageServices languageServices)
-        => languageServices.GetRequiredService<ICodeGenerationService>().GetCodeGenerationOptions(options);
-
-    public static CodeAndImportGenerationOptions GetCodeAndImportGenerationOptions(this IOptionsReader options, LanguageServices languageServices, bool? allowImportsInHiddenRegions = null)
-        => new()
-        {
-            GenerationOptions = options.GetCodeGenerationOptions(languageServices),
-            AddImportOptions = options.GetAddImportPlacementOptions(languageServices, allowImportsInHiddenRegions)
-        };
-
-    public static CleanCodeGenerationOptions GetCleanCodeGenerationOptions(this IOptionsReader options, LanguageServices languageServices, bool? allowImportsInHiddenRegions = null)
-        => new()
-        {
-            GenerationOptions = options.GetCodeGenerationOptions(languageServices),
-            CleanupOptions = options.GetCodeCleanupOptions(languageServices, allowImportsInHiddenRegions)
-        };
-
-    public static async ValueTask<CodeGenerationOptions> GetCodeGenerationOptionsAsync(this Document document, CancellationToken cancellationToken)
-    {
-        var configOptions = await document.GetAnalyzerConfigOptionsAsync(cancellationToken).ConfigureAwait(false);
-        return configOptions.GetCodeGenerationOptions(document.Project.Services);
-    }
-
-    public static async ValueTask<CodeGenerationContextInfo> GetCodeGenerationInfoAsync(this Document document, CodeGenerationContext context, CancellationToken cancellationToken)
-    {
-        Contract.ThrowIfNull(document.Project.ParseOptions);
-
-        var options = await GetCodeGenerationOptionsAsync(document, cancellationToken).ConfigureAwait(false);
-        var service = document.Project.Services.GetRequiredService<ICodeGenerationService>();
-        return service.GetInfo(context, options, document.Project.ParseOptions);
-    }
-#endif
-}
