+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/Formatting/ISyntaxFormattingService.cs
